Fix discrepancies in run time limit representation

+++ include/Forges/FuzzyHammerer.hpp
@@ -32,7 +32,7 @@ class FuzzyHammerer {
   static void
   n_sided_frequency_based_hammering(BlacksmithConfig &config, DramAnalyzer &dramAnalyzer, Memory &memory,
                                     uint64_t acts, bool fixed_acts_per_ref,
-                                    unsigned long runtime_limit, size_t probes_per_pattern, bool sweep_best_pattern);
+                                    size_t runtime_limit, size_t probes_per_pattern, bool sweep_best_pattern);
 
   static void test_location_dependence(ReplayingHammerer &rh, HammeringPattern &pattern);
